Fix: Limit laser power to available values
Summary:
Credit to @brent113 for finding this originally.

The issue is that we are allowed to call `ToolHeadLaser::SetPower` with
any value, and it will attempt to do the conversion through the
`power_table_` using that value.

This means that calling `ToolHeadLaser::SetPower` with any value above
100 will result in reading random data and attempting to set that as the
PWM value.

To fix, we should clamp `power` to the size of the `power_table_` when
doing any lookups.

Test Plan:
`pio run`
I also wrote a test of sorts (http://tpcg.io/AMS3HP), but it'd be better
if I could write a proper test for this method.