Fix: Disable inline power in TurnOff
When a job is aborted in the touchscreen, the HMI calls `TurnOff` directly in `SystemService::PreProcessStop`. However, when `TurnOff` is called, there could still be blocks in the planner, and the planner will continue to execute those blocks after `TurnOff` was called.

This ends up causing the issue that the laser can be turned back on by the planner finishing the next block, and there's nothing that will disable the laser afterwards.

If we add a `InlineDisable` into `TurnOff`, we can be sure to not turn the laser back on as the planner finishes the block. We can also remove the calls to `InlineDisable()` in the various gcode methods since they will now have `InlineDisable()` called as a part of `TurnOff()`.

Test plan:
See issue #221 for reproduction steps.

Before, the laser would potentially stay on if cancelled while an inline power block was in the queue.
After, the laser properly shuts off even if there's inline power blocks in the queue.