Fix: Disable inline power in TurnOff
When @brent113 wrote the inline controls, the M5 gcode would be executed when stopping a job through the HMI.

This doesn't appear to be the case anymore, and the HMI calls `TurnOff` directly in `SystemService::PreProcessStop`. However, when `TurnOff` is called, there could still be blocks in the planner, and the planner will continue to execute those blocks after `TurnOff` was called.

This ends up causing the issue that the laser can be turned back on by the planner finishing the next block, and there's nothing that will disable the laser afterwards.

If we add a `InlineDisable` into `TurnOff`, we can be sure to not turn the laser back on as the planner finishes the block.

Test plan:
See issue for reproduction steps: https://github.com/Snapmaker/Snapmaker2-Controller/issues/221

Before, the laser would potentially stay on if cancelled while an inline power block was in the queue.
After, the laser properly shuts off even if there's inline power blocks in the queue.