Improvement: Set system status to started when print timer started (M75)
Add an event to be sent to `systemservice` when a job is marked as started with M75.

Motivation: Currently, when a print is started from octoprint (or anything not the HMI), the `SystemService.cur_status_` is not updated. This then causes us to not handle filament runout, since there's a conditional check that the status must be `SYSTAT_WORK`: https://github.com/Snapmaker/Snapmaker2-Controller/blob/main/Marlin/src/feature/runout.h#L100

Additionally, M75 and M77 are currently "unsupported". I'm assuming that this is just that they are not actually upgraded to work with Snapmaker's cusotmiztions.

We can use M75 and M77 to update the system service status so that runout detection is activated when a print job is "started".

This diff will add the necessary `SystemService` calls to make sure that `SystemService` status is set to `SYSTAT_WORK` when a job is in progress (after M75) and back to `SYSTAT_IDLE` when the job finishes.

By itself, this will have no influence on the operation of the machine, since luban and the HMI do not ever use M75 or M77. However, in combination with some configuration in octoprint, this diff enables the ability for runout detection to be triggered when printing via octoprint.